First attempt to mark up Olle’s essays
Olle leaves essays to be read. z3998 doesn’t have any specific essay markup, and the first one at least can be easily massaged into the existing letter format, so let’s go with that in the absence of anything better.

I’m not a big fan of all the nested quotation marks but this at least mirrors the original text. If we want to change them I’m not going to argue.